:truck: chore: organize controllers' files by entity
